+++ l10n_us_sales_tax_engine/models/__init__.py
@@ -1,6 +1,7 @@
 # Copyright 2026 Binhex - Carlos R. Rodriguez.
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl).
-from . import res_config_settings  # noqa: F401
+from . import us_tax_auto_mixin  # noqa: F401
+from . import us_tax_line_hash_mixin  # noqa: F401
 from . import us_tax_jurisdiction  # noqa: F401
 from . import us_tax_zip_mapping  # noqa: F401
 from . import us_tax_rate  # noqa: F401
@@ -14,4 +15,6 @@ from . import us_tax_import_batch  # noqa: F401
 from . import res_partner  # noqa: F401
 from . import product_template  # noqa: F401
 from . import sale_order  # noqa: F401
+from . import sale_order_line  # noqa: F401
 from . import account_move  # noqa: F401
+from . import account_move_line  # noqa: F401